private/private-4-3-1/requirements.txt
anyio==3.6.2
base58==2.1.1
certifi==2023.5.7
Deprecated==1.2.13
ECPy==1.2.5
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
pycryptodome==3.18.0
rfc3986==1.5.0
sniffio==1.3.0
types-Deprecated==1.2.9.2
typing_extensions==4.6.2
websockets==10.4
wrapt==1.15.0
xrpl-py==1.8.0
